drf_practice_01/pyproject.toml
[project]
name = "drf-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.9"
dependencies = [
    "django>=6.0",
    "djangorestframework>=3.16.1",
    "python-dotenv>=1.2.1",
]
